droidflow/requirements.txt
flask
flask-socketio
apscheduler
uiautomator2
eventlet
websockets
av==10.0.0
pillow